refactor: reduce db service
