+++ src/artist/artist.service.ts
@@ -1,28 +1,40 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { DbService } from 'src/db/db.service';
 import { ArtistDto } from 'src/db/dto';
+import { Artist } from 'src/db/types';
 
 @Injectable()
 export class ArtistService {
   constructor(private readonly dbService: DbService) {}
 
   getArtists() {
-    return this.dbService.getArtists();
+    return this.dbService.artists;
   }
 
   getArtist(id: string) {
-    return this.dbService.getArtist(id);
+    const artist = this.dbService.artists.find((a) => a.id === id);
+    if (!artist) throw new NotFoundException();
+    return artist;
   }
 
   createArtist(dto: ArtistDto) {
-    this.dbService.createArtist(dto);
+    const artist: Artist = {
+      id: crypto.randomUUID(),
+      name: dto.name,
+      grammy: dto.grammy,
+    };
+    this.dbService.artists.push(artist);
   }
 
   updateArtist(id: string, dto: Partial<ArtistDto>) {
-    return this.dbService.updateArtist(id, dto);
+    const artist = this.dbService.artists.find((a) => a.id === id);
+    if (!artist) throw new NotFoundException();
+    Object.assign(artist, dto);
   }
 
   deleteArtist(id: string) {
-    this.dbService.deleteArtist(id);
+    const artistIndex = this.dbService.artists.findIndex((a) => a.id === id);
+    if (artistIndex === -1) throw new NotFoundException();
+    this.dbService.artists.splice(artistIndex, 1);
   }
 }
